+++ system/src/Grav/Common/Twig/TwigExtension.php
@@ -99,7 +99,7 @@ class TwigExtension extends \Twig_Extension
             new \Twig_SimpleFilter('json_decode', [$this, 'jsonDecodeFilter']),
             new \Twig_SimpleFilter('array_unique', 'array_unique'),
             new \Twig_SimpleFilter('basename', 'basenameFilter'),
-
+            new \Twig_SimpleFilter('dirname', 'dirnameFilter'),
         ];
     }
 
@@ -1124,5 +1124,10 @@ class TwigExtension extends \Twig_Extension
     public function basenameFilter($var)
     {
         return basename($var);
+    }
+
+    public function dirnameFilter($var)
+    {
+        return dirname($var);
     }
 }